Refactor appStateChange listener to hook
